requirements.txt
streamlit
smartapi-python
pyotp
logzero
websocket-client
requests
pandas
pytz
